requirements.txt
pymilvus 
python-dotenv
torch
transformers
langchain-text-splitters